requirements.txt
attrs==21.2.0
charset-normalizer==2.0.7
configobj==5.0.6
coverage==6.1.2
DateTime==4.3
mysql-connector-python==8.0.27
numpy==1.21.4
packaging==21.2
peewee==3.14.8
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
rand==0.3.0
six==1.16.0
